Add infrastructure for reminders
Move stuff for date and number parsing from todo to reminders.
Move file access to fileHandler.

Add and remove reminders. Automatically add new and remove old when
adding or changing due date in todo list.
