+++ Jarvis/fileHandler.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+import os
+import json
+from datetime import datetime as dt
+import uuid
+
+timeFormat = "%Y-%m-%d %H:%M:%S"
+
+def json_serial(obj):
+    """JSON serializer for objects not serializable by default json code"""
+
+    if isinstance(obj, dt):
+        serial = obj.strftime(timeFormat)
+        return serial
+    elif isinstance(obj, uuid.UUID):
+        serial = obj.hex
+        return serial
+    raise TypeError ("Type not serializable")
+
+def readFile(name, default = []):
+    if os.path.exists(name):
+        with open(name, "r+") as f:
+            return json.load(f)
+    return default
+
+def writeFile(name, obj):
+    with open(name, "w+") as f:
+        json.dump(obj, f, default=json_serial)
+
+def str2date(string):
+    return dt.strptime(string, timeFormat)
