+++ Jarvis/todo.py
@@ -4,24 +4,13 @@ import sys
 import json
 
 from datetime import datetime as dt
-from datetime import date, time, timedelta
-from dateutil.relativedelta import relativedelta
-import re
+import uuid
+
+from reminder import parseNumber, parseDate, addReminder, removeReminder
+from fileHandler import writeFile, readFile, str2date
 
 from colorama import init
 from colorama import Fore, Back, Style
-
-def json_serial(obj):
-    """JSON serializer for objects not serializable by default json code"""
-
-    if isinstance(obj, dt):
-        serial = obj.strftime("%Y-%m-%d %H:%M:%S")
-        return serial
-    raise TypeError ("Type not serializable")
-
-def writeToFile():
-    with open("todolist.txt", "w+") as f:
-        json.dump(todoList, f, default=json_serial)
 
 def printItem(item, index):
     if 'priority' in item and item['priority'] >= 50:
@@ -29,7 +18,11 @@ def printItem(item, index):
     if 'priority' in item and item['priority'] >= 100:
         sys.stdout.write(Fore.RED)
     print("<{2}> {0} [{1}%]".format(item['name'], item['complete'], index) + Fore.RESET)
+    if not isinstance(item['uuid'], uuid.UUID):
+        item['uuid'] = uuid.UUID(item['uuid'])
     if 'due' in item:
+        if not isinstance(item['due'], dt):
+            item['due'] = str2date(item['due'])
         if item['due'] < dt.now():
             sys.stdout.write(Fore.RED)
         print("\tDue at {0}".format(item['due'].strftime("%Y-%m-%d %H:%M:%S")) + Fore.RESET)
@@ -49,118 +42,6 @@ def sort(data):
             l['items'] = sort(l['items'])
     return sorted(data, key = lambda k: (-k['priority'] if 'priority' in k else 0, k['complete']))
 
-def parseNumber(string, numwords = {}):
-    if not numwords:
-        units = ["zero", "one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten", "eleven", "twelve", "thirteen", "fourteen", "fifteen", "sixteen", "seventeen", "eighteen", "nineteen" ]
-        tens = ["", "", "twenty", "thirty", "fourty", "fifty", "sixty", "seventy", "eighty", "ninety"]
-        scales = ["hundred", "thousand", "million", "billion", "trillion"]
-
-        numwords["and"] = (1, 0)
-        for idx, word in enumerate(units):    numwords[word] = (1, idx)
-        for idx, word in enumerate(tens):     numwords[word] = (1, idx * 10)
-        for idx, word in enumerate(scales):   numwords[word] = (10 ** (idx * 3 or 2), 0)
-
-    ret = {'skip':0, 'value':0}
-    elements = string.replace(",", "").split()
-    current = 0
-    for d in elements:
-        number = d.split("-")
-        for word in number:
-            if word not in numwords:
-                try:
-                    scale, increment = (1, int(word))
-                except ValueError:
-                    ret['value'] += current
-                    return ret
-            else:
-                scale, increment = numwords[word]
-                if not current and scale > 100:
-                    current = 1
-            current = current * scale + increment
-            if scale > 100:
-                ret['value'] += current
-                current = 0
-        ret['skip'] += 1
-    ret['value'] += current
-    return ret
-    
-def parseDate(data):
-    elements = data.split()
-
-    parseDay = False
-    parseDeltaValue = False
-    parseDeltaUnit = 0
-    deltaValue = 0
-    retDate = dt.now().date()
-    retTime = time(23, 59, 59)
-    for index, d in enumerate(elements):
-        if parseDay:
-            d += dt.today().strftime(" %Y %W")
-            try:
-                retDate = dt.strptime(d, "%a %Y %W").date()
-            except ValueError:
-                try:
-                    retDate = dt.strptime(d, "%A %Y %W").date()
-                except ValueError:
-                    print("Could not parse word: {0}".format(d))
-                    parseDay = False
-                    continue
-            if retDate <= dt.now().date():
-                retDate += timedelta(days = 7)
-            parseDay = False
-        elif parseDeltaValue:
-            tmp = parseNumber(" ".join(elements[index:]))
-            deltaValue = tmp['value']
-            parseDeltaUnit = tmp['skip']
-            parseDeltaValue = False
-        elif parseDeltaUnit:
-            if "year" in d:
-                retDate += relativedelta(years = deltaValue)
-            elif "month" in d:
-                retDate += relativedelta(months = deltaValue)
-            elif "week" in d:
-                retDate += timedelta(weeks = deltaValue)
-            elif "day" in d:
-                retDate += timedelta(days = deltaValue)
-            elif "hour" in d:
-                newTime = dt.now() + timedelta(hours = deltaValue)
-                retDate = newTime.date()
-                retTime = newTime.time()
-            elif "minute" in d:
-                newTime = dt.now() + timedelta(minutes = deltaValue)
-                retDate = newTime.date()
-                retTime = newTime.time()
-            elif "second" in d:
-                newTime = dt.now() + timedelta(seconds = deltaValue)
-                retDate = newTime.date()
-                retTime = newTime.time()
-            elif parseDeltaUnit == 1:
-                print("Missing time unit")
-            parseDeltaUnit -= 1
-
-        elif re.match("^[0-9]{2}-[0-1][0-9]-[0-3][0-9]", d):
-            retDate = dt.strptime(d, "%y-%m-%d").date()
-        elif re.match("^[1-9][0-9]{3}-[0-1][0-9]-[0-3][0-9]", d):
-            retDate = dt.strptime(d, "%Y-%m-%d").date()
-        elif re.match("^[0-3][0-9]-[0-1][0-9]-[0-9]{2}", d):
-            pretDate = dt.strptime(d, "%d.%m.%y").date()
-        elif re.match("^[0-3][0-9]-[0-1][0-9]-[1-9][0-9]{3}", d):
-            retDate = dt.strptime(d, "%d.%m.%Y").date()
-
-        elif re.match("^[0-1][0-9]:[0-5][0-9][AP]M", d):
-            retTime = dt.strptime(d, "%I:%M%p").time
-        elif re.match("^[0-2][0-9]:[0-5][0-9]", d):
-            retTime = dt.strptime(d, "%H:%M").time
-
-        elif d == "next":
-            parseDay = True
-        elif d == "in":
-            parseDeltaValue = True
-        else:
-            print("Unknown Format: {0}".format(d))
-            continue
-    return dt.combine(retDate, retTime)
-
 def getItem(string, todoList):
     words = string.split(".")
     retList = []
@@ -176,7 +57,6 @@ def getItem(string, todoList):
     return retList
 
 def todoHandler(data):
-    global todoList
     if "add" in data:
         data = data.replace("add", "", 1)
         if "comment" in data:
@@ -195,7 +75,9 @@ def todoHandler(data):
             item = todoList
             for i in index:
                 item = item['items'][i]
+            removeReminder(item['uuid'].hex)
             item['due'] = parseDate(" ".join(words[1:]))
+            addReminder(name=item['uuid'].hex, time=item['due'])
         else:
             data = " ".join(data.split())
             try:
@@ -206,7 +88,7 @@ def todoHandler(data):
                 data = " ".join(data.split()[1:])
             except ValueError:
                 item = todoList
-            newItem = {'complete':0}
+            newItem = {'complete':0, 'uuid':uuid.uuid4()}
             parts = data.split(" - ")
             newItem['name'] = parts[0]
             if " - " in data:
@@ -259,17 +141,8 @@ def todoHandler(data):
 
     todoList['items'] = sort(todoList['items'])
     _print(todoList['items'])
-    writeToFile()
+    writeFile("todolist.txt", todoList)
 
-todoList = {}
-todoList['items'] = []
-if not os.path.exists("todolist.txt"):
-    writeToFile()
-else:
-    with open("todolist.txt", "r+") as f:
-        todoList = json.load(f)
-        todoList['items'] = sort(todoList['items'])
-        for i in todoList['items']:
-            if 'due' in i:
-                i['due'] = dt.strptime(i['due'], "%Y-%m-%d %H:%M:%S")
+todoList = readFile("todolist.txt", {'items':[]})
+todoList['items'] = sort(todoList['items'])
